improve test_gtcpp_compilation
